+++ kosta_finalpjt/src/main/java/com/example/demo/depts/DeptsDao.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.demo.member.Members;
+import com.example.demo.members.Members;
 
 @Repository
 public interface DeptsDao extends JpaRepository<Depts, Integer> {
