+++ kosta_finalpjt/src/main/java/com/example/demo/depts/DeptsService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.example.demo.member.Members;
+import com.example.demo.members.Members;
 
 @Service
 public class DeptsService {
